backend/requirements.txt
asgiref==3.3.4
coverage==5.5
django==3.0.7
django-cors-headers==3.7.0
djangorestframework==3.10
pytz==2021.1
sqlparse==0.4.1
gunicorn==20.1.0
